[IMP] account_analytic_required: do not enforce policy when credit=debit=0.0
This is useful in some situations where the tax generates move lines with credit=debit=0.0 with a non-zero tax amount (for instance with tax code VAT-IN-V82-21-EU-S in l10n_be). Without this patch the transaction would be rejected when analytic policy is 'required', since analytic accounts are not propagated through taxes in 6.1.