[FIX] analytic_tag_dimension: avoid column not exists
Creating dimensions, you can face this error:

psycopg2.ProgrammingError: column ail.x_dimension_xxx does not exist
LINE 48:         , ail.x_dimension_xxx as x_dimension_xxx, ai.team_..

This is because the column is not yet populated in the moment when
the model initialization is triggered.

In this commit, we inhibit the population of that field in the report
on that case, and launching manual initialization later when done.