[IMP] analytic_tag_dimension: Don't write 2 times when using tags
Previous code was doing a deferred write after the creation/first write for easing the
code, but it's very bad for performance in such used objects.

This commits changes the way this is handled for covering the cases in advance, and thus
not requiring a second write. The only drawback is that writing on multiple records
means to perform a write on each record separately. This case was also happening
previously, and we have gained one write (the initial one), but it's not the ideal one
in terms of performance. A possible improvement can be to group the records with the
same final values to be performed at once.