[FIX] analytic_tag_dimension: Avoid non created fields when model is populated after the dimension creation
If you create a dimension (for example, through demo data), and after
this a new module inheriting and changing name from one of the affected
models (like hr.timesheet.switch model inheriting from
account.analytic.line in project_timesheet_time_control module), the
dimension fields are not created, obtaining an error if you for example
do a read.

We intercept here the `_reflect` method that is called when initializing
the model for injecting the same fields in this phase.

TT29728