[IMP] Add config setting to always use product analytic
Use case:
We have a service product, invoiced on timesheet and that triggers the creation of a project when a SO is confirmed. Note that doing this, an analytic account is created and linked to the SO (module sale_project)
An income analytic account is defined on this product. We want this analytic account to be in the analytic distribution of the invoice lines that are created from the SO.
Now, this is not possible. When creating an invoice, the analytic distribution on the invoice lines is the SO analytic account.
This option forces the use of the product analytic account on the invoice line instead.